fix sql

+++ src/bot.rs
@@ -19,7 +19,7 @@ macro_rules! reply_i18n_and_return {
 #[macro_export]
 macro_rules! parse_integer {
     ($bot:expr, $chat_id:expr, $integer:expr) => {{
-        let out: i64 = match $integer.parse() {
+        let out: i32 = match $integer.parse() {
             Ok(integer) => integer,
             Err(_) => {
                 $bot.send_message($chat_id, t!("not_valid_integer")).await?;
